Move UpdateItemQuantityRequest to cart package
